Update Calculator and Create new child component
