+++ src/components/Button.js
@@ -1,0 +1,14 @@
+import PropTypes from 'prop-types';
+
+const Button = ({ type, label }) => (
+  <button type="button" className={`button ${type}`}>
+    {label}
+  </button>
+);
+
+Button.propTypes = {
+  type: PropTypes.string.isRequired,
+  label: PropTypes.string.isRequired,
+};
+
+export default Button;
