+++ src/components/Calculator.js
@@ -1,16 +1,5 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-
-const Button = ({ type, label }) => (
-  <button type="button" className={`button ${type}`}>
-    {label}
-  </button>
-);
-
-Button.propTypes = {
-  type: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-};
+import Button from './Button';
 
 const Calculator = () => (
   <div className="calculator">
